basic v3/v4 js/ts functions
