+++ api-functions-v3-js/package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "",
   "scripts": {
+    "preinstall": "npm install -g azure-functions-core-tools@4 --unsafe-perm true",
     "start": "func start",
     "test": "echo \"No tests yet...\""
   },
