+++ api-functions-v3-js/status/index.js
@@ -1,0 +1,8 @@
+module.exports = async function (context, req) {
+    context.log('JavaScript HTTP trigger function processed a request.');
+
+    context.res = {
+        // status: 200, /* Defaults to 200 */
+        body: process.env
+    };
+}
