+++ api-functions-v3/status/index.ts
@@ -1,0 +1,13 @@
+import { AzureFunction, Context, HttpRequest } from "@azure/functions"
+
+const httpTrigger: AzureFunction = async function (context: Context, req: HttpRequest): Promise<void> {
+    context.log('HTTP trigger function processed a request.');
+
+    context.res = {
+        // status: 200, /* Defaults to 200 */
+        body: process.env
+    };
+
+};
+
+export default httpTrigger;
