+++ api-functions-v4-js/src/functions/status.js
@@ -6,8 +6,7 @@ async function status(request, context) {
     return {
         status: 200,
         jsonBody: {
-            env: process.env,
-            requestHeaders: request.headers
+            env: process.env
         }
     };
 };
