+++ api-functions-v4/package.json
@@ -3,15 +3,19 @@
   "version": "1.0.0",
   "description": "",
   "scripts": {
+    "preinstall": "npm install -g azure-functions-core-tools@4 --unsafe-perm true",
     "build": "tsc",
     "watch": "tsc -w",
     "prestart": "npm run build",
     "start": "func start",
     "test": "echo \"No tests yet...\""
   },
+  "dependencies": {
+    "@azure/functions": "^4.0.0-alpha.7"
+  },
   "devDependencies": {
-    "@azure/functions": "^3.0.0",
-    "@types/node": "^16.x",
+    "@types/node": "^18.x",
     "typescript": "^4.0.0"
-  }
+  },
+  "main": "dist/src/functions/*.js"
 }
